+++ enro-core/src/main/java/dev/enro/core/controller/createNavigationController.kt
@@ -2,13 +2,17 @@ package dev.enro.core.controller
 
 import android.app.Application
 import androidx.annotation.Keep
+import dev.enro.core.EnroConfig
 
 /**
  * Create a NavigationController from the NavigationControllerDefinition/DSL, and immediately attach it
  * to the NavigationApplication from which this function was called.
+ *
+ * @param useLegacyContainerPresentBehavior see [EnroConfig.useLegacyContainerPresentBehavior]
  */
 public fun NavigationApplication.createNavigationController(
     strictMode: Boolean = false,
+    useLegacyContainerPresentBehavior: Boolean = false,
     backConfiguration: EnroBackConfiguration = EnroBackConfiguration.Default,
     block: NavigationModuleScope.() -> Unit = {}
 ): NavigationController {
@@ -22,6 +26,7 @@ public fun NavigationApplication.createNavigationController(
         setConfig(
             config.copy(
                 isStrictMode = strictMode,
+                useLegacyContainerPresentBehavior = useLegacyContainerPresentBehavior,
                 backConfiguration = backConfiguration,
             )
         )
@@ -35,14 +40,21 @@ public fun NavigationApplication.createNavigationController(
 )
 public fun NavigationApplication.navigationController(
     strictMode: Boolean = false,
+    useLegacyContainerPresentBehavior: Boolean = false,
     backConfiguration: EnroBackConfiguration = EnroBackConfiguration.Default,
     block: NavigationModuleScope.() -> Unit = {}
-): NavigationController = createNavigationController(strictMode, backConfiguration, block)
+): NavigationController = createNavigationController(
+    strictMode = strictMode,
+    useLegacyContainerPresentBehavior = useLegacyContainerPresentBehavior,
+    backConfiguration = backConfiguration,
+    block = block
+)
 
 
 @Keep // Used by EnroTest
 internal fun createUnattachedNavigationController(
     strictMode: Boolean = false,
+    useLegacyContainerPresentBehavior: Boolean = false,
     backConfiguration: EnroBackConfiguration = EnroBackConfiguration.Default,
     block: NavigationModuleScope.() -> Unit = {}
 ): NavigationController {
@@ -52,6 +64,7 @@ internal fun createUnattachedNavigationController(
         setConfig(
             config.copy(
                 isStrictMode = strictMode,
+                useLegacyContainerPresentBehavior = useLegacyContainerPresentBehavior,
                 backConfiguration = backConfiguration,
             )
         )